pyproject.toml
[project]
name = "over-particular"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pygame-ce>=2.5.6",
    "typing>=3.10.0.0",
]
